+++ tests/data/test_import_affymetrix.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Jun  9 15:44:58 2021
+
+@author: Paolo Cozzi <paolo.cozzi@ibba.cnr.it>
+"""